# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Truck App API",
        "description": "API Documentation for Truck App Backend",
        "contact": {
            "email": "dabeer445@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Truck App Backend"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve a list of orders",
                "operationId": "6390e58970c6b576bd8c1ec198cb5fbb",
                "responses": {
                    "200": {
                        "description": "A list of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderCollection"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "8cb746eb001b58ac915930ba43cfdf0d",
                "requestBody": {
                    "description": "Order creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve a specific order",
                "operationId": "7830784c8e02a11ee0de8d1e6c90a0fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update a specific order",
                "operationId": "6fe2d9225419ef1f17fb5c3270fc92bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pickup_location",
                                    "delivery_location",
                                    "cargo_details",
                                    "pickup_time",
                                    "delivery_time"
                                ],
                                "properties": {
                                    "pickup_location": {
                                        "type": "string"
                                    },
                                    "delivery_location": {
                                        "type": "string"
                                    },
                                    "cargo_details": {
                                        "properties": {
                                            "weight": {
                                                "type": "number"
                                            },
                                            "dimensions": {
                                                "properties": {
                                                    "length": {
                                                        "type": "number"
                                                    },
                                                    "width": {
                                                        "type": "number"
                                                    },
                                                    "height": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "pickup_time": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "delivery_time": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete a specific order",
                "operationId": "75a225eace7e4d30255b5d45e05c41fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}/cancel": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel a specific order",
                "operationId": "ff31150366af592be674225bd2df113f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Only pending orders can be cancelled"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    }
                },
                "type": "object"
            },
            "OrderRequest": {
                "required": [
                    "pickup_location",
                    "delivery_location",
                    "cargo_details",
                    "pickup_time",
                    "delivery_time"
                ],
                "properties": {
                    "pickup_location": {
                        "type": "string",
                        "example": "123 Pickup St"
                    },
                    "delivery_location": {
                        "type": "string",
                        "example": "456 Delivery Ave"
                    },
                    "cargo_details": {
                        "type": "object",
                        "example": {
                            "weight": "50kg",
                            "dimensions": "100x50x75cm"
                        }
                    },
                    "pickup_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "delivery_time": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Order": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/OrderRequest"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "user_id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "pending",
                                    "in_progress",
                                    "completed",
                                    "cancelled"
                                ],
                                "example": "pending"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "deleted_at": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "OrderCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "from": {
                                "type": "integer"
                            },
                            "last_page": {
                                "type": "integer"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CreateOrderRequest": {
                "required": [
                    "pickup_location",
                    "delivery_location",
                    "cargo_details",
                    "pickup_time",
                    "delivery_time"
                ],
                "properties": {
                    "pickup_location": {
                        "type": "string",
                        "example": "123 Pickup St"
                    },
                    "delivery_location": {
                        "type": "string",
                        "example": "456 Delivery Ave"
                    },
                    "cargo_details": {
                        "required": [
                            "weight",
                            "dimensions"
                        ],
                        "properties": {
                            "weight": {
                                "type": "number",
                                "format": "float",
                                "maximum": 5000,
                                "minimum": 1,
                                "example": 100
                            },
                            "dimensions": {
                                "required": [
                                    "length",
                                    "width",
                                    "height"
                                ],
                                "properties": {
                                    "length": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 1,
                                        "example": 10
                                    },
                                    "width": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 1,
                                        "example": 10
                                    },
                                    "height": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 1,
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "pickup_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T10:00:00Z"
                    },
                    "delivery_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-02T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}